#457 FigureApi, ObjectApi methods moved
