+++ supervisely_lib/api/pointcloud/pointcloud_figure_api.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 from supervisely_lib.video_annotation.key_id_map import KeyIdMap
-from supervisely_lib.api.entity_annotation.figure_api import FigureApi
+from supervisely_lib.api.entity_annotation.figure_api import FigureApi, ApiField
 
 
 class PointcloudFigureApi(FigureApi):
@@ -17,6 +17,17 @@ class PointcloudFigureApi(FigureApi):
 
         self._append_bulk(pointcloud_id, figures_json, keys, key_id_map)
 
+    def append_to_dataset(self, dataset_id, figures, entity_ids, key_id_map: KeyIdMap):
+        keys = []
+        figures_json = []
+        for figure, entity_id in zip(figures, entity_ids):
+            keys.append(figure.key())
+            figure_json = figure.to_json(key_id_map)
+            figure_json[ApiField.ENTITY_ID] = entity_id
+            figures_json.append(figure_json)
+
+        return self._append_bulk(dataset_id, figures_json, keys, key_id_map, field_name=ApiField.DATASET_ID)
+
     def _convert_json_info(self, info: dict, skip_missing=True):
-                return super()._convert_json_info(info, skip_missing)
+        return super()._convert_json_info(info, skip_missing)
