+++ supervisely_lib/api/pointcloud/pointcloud_object_api.py
@@ -11,4 +11,5 @@ class PointcloudObjectApi(ObjectApi):
         return self._append_bulk(self._api.pointcloud.tag, pointcloud_id, info.project_id, info.dataset_id, objects, key_id_map, is_pointcloud=True)
 
     def append_to_dataset(self, dataset_id, objects: PointcloudObjectCollection, key_id_map: KeyIdMap = None):
-        return self._append_to_dataset(self._api.pointcloud.tag, dataset_id, objects, key_id_map, is_pointcloud=True)
+        project_id = self._api.dataset.get_info_by_id(dataset_id).project_id
+        return self._append_bulk(self._api.pointcloud.tag, dataset_id, project_id, dataset_id, objects, key_id_map, is_pointcloud=True)
